Open that stuff. You know you want to.
Adds opening of files on double click and changes file preview to single click. Most of the work done by Tom, I just added the conversion of the path from relative to view to absolute and changed the lambda to use that instead.
Also got rid of an annoying untagged debug output.